
\name{fixed}
\Rdversion{1.0}
\alias{fixed}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
  Extracting and Summarizing Fixed Effects
}

\description{
This function is to extract the estimates, standard deviations and p-values of fixed effects.
}

\usage{
fixed(object) 
fixed.mms(object)
fixed.mgam(object)
fixed.mglmmTMB(object)
}

%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{ 
  an object from \code{\link{mms}}, \code{\link{mgam}, or \code{\link{mglmmTMB}}}.
  }
}

\details{

}

\value{
  These functions return the estimates, standard deviations and p-values of fixed effects.
}

\author{
  Nengjun Yi, nyi@uab.edu
}

\seealso{

}

\examples{
see examples in \code{\link{mms}, \code{\link{mgam}, or \code{\link{mglmmTMB}}}} 
}

