
\name{NBZIMM-package}
\Rdversion{1.0}
\alias{NBZIMM-package}
\alias{NBZIMM}
\docType{package}
\title{
  Negative Binomial and Zero-Inflated Mixed Models
}

\description{
  This package provides functions for setting up and fitting negative binomial mixed models and zero-inflated negative binomial and Gaussian mixed models. These functions allow for multiple and correlated group-specific (random) effects and various types of within-group correlation structures as described in the core package \bold{nlme}, and return objects that can be summarized by functions in \bold{nlme}. The methods can be used to analyze overdispersed and zero-inflated count or continuous responses with multilevel data structures (for example, clustered data and longitudinal studies).   
}

\details{
  \tabular{ll}{
  Package: \tab NBZIMM\cr
  Type: \tab Package\cr
  Version: \tab 1.0\cr
  Date: \tab 2024-06-11\cr
  License: \tab MIT\cr
  LazyLoad: \tab yes\cr
  }
}

\author{
  Nengjun Yi, nyi@uab.edu
}


